+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [unreleased]
 
 ### Added
-
 - Added the anchor to the crop insurance county level maps and reuseable function for the county-level maps in the future [#445](https://github.com/policy-design-lab/pdl-frontend/issues/445)
+
+### Changed
+- Upgrade ESLint to version 9 and adjusted code to meet the new standards [#447](https://github.com/policy-design-lab/pdl-frontend/issues/447)
 
 ## [1.15.0] - 2026-03-04